+++ guizero/ButtonGroup.py
@@ -14,7 +14,7 @@ class ButtonGroup(
     DisplayMixin, 
     ReprMixin):
 
-    def __init__(self, master, options, selected, horizontal=False, command=None, grid=None, align=None, args = None):
+    def __init__(self, master, options, selected, horizontal=False, command=None, grid=None, align=None, args=None):
         
         self._master = master
         self._grid = grid
@@ -58,6 +58,9 @@ class ButtonGroup(
                 gridx += 1
             else:
                 gridy += 1
+
+            # Set the callback
+            rbutton.tk.config(command=self._command_callback)
 
         # Add a command if there was one
         self.update_command(command, args)
@@ -108,18 +111,17 @@ class ButtonGroup(
     def get_group_as_list(self):
         return [[option.text, option.value] for option in self._options]
 
-    def update_command(self, newcommand, args=None):
-        if newcommand is None:
+    def update_command(self, command, args=None):
+        if command is None:
             self._command = lambda: None
         else:
             if args is None:
-                self._command = newcommand
+                self._command = command
             else:
-                # If the args LIST was not blank, allow args
-                self._command = utils.with_args(newcommand, *args)
-        
-        for rbutton in self._options:
-            rbutton.tk.config(command=self._command)
+                self._command = utils.with_args(command, *args)
+    
+    def _command_callback(self):
+        self._command()
 
     # DEPRECATED METHODS
     # -----------------------------------