+++ guizero/PushButton.py
@@ -24,13 +24,13 @@ class PushButton(
         self._current_font = "Arial"
         self._font_size = 11
         self._value = 0
-        self.change_command(command, args)
+        self.update_command(command, args)
         
         # Description of this object (for friendly error messages)
         self.description = "[PushButton] object with text \"" + self._text.get() + "\""
 
         # Create a tk Button object within this object
-        self.tk = Button(master.tk, textvariable=self._text)
+        self.tk = Button(master.tk, textvariable=self._text, command=self._command_callback)
 
         # Add padding if necessary
         self.tk.config(pady=pady, padx=padx)
@@ -147,19 +147,7 @@ class PushButton(
     def _on_release(self, event):
         if self.enabled:
             self._value = 0
-            self._command()
         
-    # Change command - needs the name of a function and optional args as a list
-    def change_command(self, newcommand, args=None):
-        if newcommand is None:
-            self._command = lambda: None
-        else:
-            if args is None:
-                self._command = newcommand
-            else:
-                # If the args LIST was not blank, allow args
-                self._command = utils.with_args(newcommand, *args)
-
     # Change padding
     def padding(self, padx, pady):
         self.tk.config(padx=padx, pady=pady)
@@ -176,6 +164,18 @@ class PushButton(
     def toggle(self):
         self.enabled = not self.enabled
 
+    def update_command(self, command, args=None):
+        if command is None:
+            self._command = lambda: None
+        else:
+            if args is None:
+                self._command = command
+            else:
+                self._command = utils.with_args(command, *args)
+
+    def _command_callback(self):
+        self._command()
+
     # DEPRECATED
     # -------------------------------------------
     # Change text
@@ -187,3 +187,8 @@ class PushButton(
     def toggle_state(self):
         self.toggle()
         utils.deprecated("PushButton toggle_state() is deprecated - renamed to toggle()")
+
+    # Change command - needs the name of a function and optional args as a list
+    def change_command(self, newcommand, args=None):
+        self.update_command(newcommand, args)
+        utils.deprecated("PushButton change_command() is deprecated - renamed to update_command()")