+++ guizero/CheckBox.py
@@ -12,7 +12,7 @@ class CheckBox(
     DisplayMixin, 
     ReprMixin):
 
-    def __init__(self, master, text, command=None, grid=None, align=None):
+    def __init__(self, master, text, command=None, grid=None, align=None, args=None):
 
         self._master = master
         self._grid = grid
@@ -26,9 +26,9 @@ class CheckBox(
         # Create a tk Checkbutton object within this object
         self.tk = Checkbutton(master.tk, text=text, variable=self._value)
 
-        # Add a command if there was one
-        if command is not None:
-            self.tk.config(command=command)
+        # Set the command callback
+        self.tk.config(command=self._command_callback)
+        self.update_command(command, args)
 
         utils.auto_pack(self, master, grid, align)
 
@@ -68,6 +68,18 @@ class CheckBox(
     def toggle(self):
         self.tk.toggle()
 
+    def update_command(self, command, args=None):
+        if command is None:
+            self._command = lambda: None
+        else:
+            if args is None:
+                self._command = command
+            else:
+                self._command = utils.with_args(command, *args)
+    
+    def _command_callback(self):
+        self._command()
+        
     # DEPRECATED METHODS
     # --------------------------------------------
     # Return text associated with this checkbox