+++ docs-src/docs/changelog.md
@@ -5,6 +5,11 @@
 - made PushButton command optional
 - Combo command functions can now have 0 arguments
 - Waffle command functions can now have 0 arguments
+- refactored command and added update_command to: 
+  - ButtonGroup
+  - CheckBox
+  - Combo
+  - PushButton
 
 ## 0.4.3 - 2017-01-10