+++ docs-src/docs/changelog.md
@@ -1,5 +1,23 @@
 # guizero
 
+## 0.4.4 - tbc
+
+- made PushButton command optional
+- Combo command functions can now have 0 arguments
+- Waffle command functions can now have 0 arguments
+- refactored command functions and added update_command to: 
+  - ButtonGroup
+  - CheckBox
+  - Combo
+  - PushButton
+  - Slider
+  - Waffle
+- refactored text / font methods and added them to:
+  - PushButton
+  - Text
+  - TextBox
+  
+
 ## 0.4.3 - 2017-01-10
 
 Minor features, bug fixes and internal refactoring